refactored codes
