+++ package.json
@@ -53,12 +53,15 @@
     "bcrypt": "^3.0.2",
     "bcrypt-node": "^0.1.0",
     "body-parser": "^1.18.3",
+    "cors": "^2.8.5",
     "cross-env": "^5.2.0",
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
+    "express-validator": "^5.3.1",
     "faker": "^4.1.0",
     "jsonwebtoken": "^8.4.0",
     "morgan": "^1.9.1",
-    "pg": "^7.7.1"
+    "pg": "^7.7.1",
+    "validator": "^10.11.0"
   }
 }
