+++ server/controllers/intervention.js
@@ -46,8 +46,8 @@ const intervention = {
 
   /**
    * Get All Intervention Incidence
-   * @param {object} req 
-   * @param {object} res 
+   * @param {object} req - HTTP Request
+   * @param {object} res - HTTP Response
    * @returns {object} incident object
    */
 
@@ -197,52 +197,6 @@ const intervention = {
         return res.status(500).json({
           status: 500,
           message: 'Server is down. Please check back in a few minutes'
-        });
-      }
-    },
-
-    /**
-     * Admin can edit intervention status
-     */
-
-    async patchInterventionStatus(req, res) {
-      const findOneQuery = 'SELECT * FROM incidents WHERE id=$1 AND type=$2';
-      const findOneAdmin = 'SELECT * FROM users WHERE id=$1 AND isAdmin=$2';
-      const patchOneQuery = `UPDATE incidents SET status=$1 WHERE id=$2 returning *`;
-
-      try {
-        const { rows } = await db.query(findOneAdmin, [req.user.id, 'true']);
-        if(rows.length === 0){
-          return res.status(400).json({
-            status: 400,
-            message: 'Not an Admin'
-          })
-        }
-        const { rows:response } = await db.query(findOneQuery, [req.params.id, 'intervention']);
-        if(!response[0]) {
-          return res.status(400).json({
-            status: 400,
-            message: 'This intervention does not exist'
-          });
-        }
-        const values = [
-          req.body.status || response[0].status,
-          req.params.id
-        ];
-        
-        const { rows:respond } = await db.query(patchOneQuery, values);
-        return res.status(200).json({
-          status: 200,
-          data: [{
-            id: respond[0].id,
-            message: 'Updated intervention record status'
-          }]
-        });
-      }
-      catch (err) {
-        return res.status(400).json({
-          status: 400,
-          message: err
         });
       }
     },
